+++ src/from/__tests__/fromVamas.test.js
@@ -8,10 +8,21 @@ describe('fromVamas', () => {
     join(__dirname, '../../../testFiles/multiplex.vms'),
     'utf8',
   );
-
+  const augerText = readFileSync(
+    join(__dirname, '../../../testFiles/auger.vms'),
+    'utf8',
+  );
   it('multiplex', () => {
     let result = fromVamas(text).spectra[0];
     expect(Object.keys(result.variables)).toHaveLength(3);
+    expect(Object.keys(result.meta)).toHaveLength(62);
+    expect(typeof result.meta.cheminfo.meta).toBe('object');
+  });
+
+
+  it('auger', () => {
+    let result = fromVamas(augerText).spectra[0];
+    expect(Object.keys(result.variables)).toHaveLength(2);
     expect(Object.keys(result.meta)).toHaveLength(62);
     expect(typeof result.meta.cheminfo.meta).toBe('object');
   });